+++ optaplanner-examples/src/main/java/org/optaplanner/examples/vehiclerouting/domain/Depot.java
@@ -46,7 +46,7 @@ public class Depot extends AbstractPersistable {
      * @param standstill never null
      * @return a positive number, the distance multiplied by 1000 to avoid floating point arithmetic rounding errors
      */
-    public int getDistanceTo(Standstill standstill) {
+    public long getDistanceTo(Standstill standstill) {
         return location.getDistanceTo(standstill.getLocation());
     }
