+++ optaplanner-examples/src/main/java/org/optaplanner/examples/vehiclerouting/domain/location/segmented/RoadSegmentLocation.java
@@ -59,10 +59,10 @@ public class RoadSegmentLocation extends Location {
     }
 
     @Override
-    public int getDistanceTo(Location location) {
+    public long getDistanceTo(Location location) {
         Double distance = getDistanceDouble((RoadSegmentLocation) location);
         // Multiplied by 1000 to avoid floating point arithmetic rounding errors
-        return (int) (distance * 1000.0 + 0.5);
+        return (long) (distance * 1000.0 + 0.5);
     }
 
     public Double getDistanceDouble(RoadSegmentLocation location) {
